Move actions to actions folder
