+++ Source/Actions/Boilerplates/installedBoilerplates.js
@@ -4,7 +4,7 @@
 *--------------------------------------------------------------------------------------------*/
 
 import path from 'path';
-import { Outputter } from '../../../Outputter';
+import { Outputter } from '../../Outputter';
 import { BoilerplatesManager } from '@dolittle/tooling.common';
